getCategoryBySlug improvement
